feat: introduce useScrollToHash hook
