+++ src/hooks/index.js
@@ -1,4 +1,5 @@
 import useDocumentTitle from './useDocumentTitle'
 import useMetaTags from './useMetaTags'
+import useScrollToHash from './useScrollToHash'
 
-export { useDocumentTitle, useMetaTags }
+export { useDocumentTitle, useMetaTags, useScrollToHash }
